* FIX [hash_table.c] fixed a logic error and a deadlock error. * NEW [test*] Add test module for hash_table. * MDF [CMakeLists.txt && Makefile] update cmakelist and makefile for test.
