+++ nanolib/include/test.h
@@ -1,0 +1,7 @@
+#include "mqtt_db.h"
+#ifndef TEST_H
+
+int dbtree_test();
+int hash_test();
+
+#endif
